🥕 v0.16.0-a.7 🌱 ThreeD Garden: Cleanup Time
