--- error-log-mm.txt
@@ -1,2 +1,0 @@
-Failed to fetch RSC payload for http://localhost:7777/participate. Falling back to browser navigation. TypeError: NetworkError when attempting to fetch resource.
-
