+++ package.json
@@ -47,6 +47,11 @@
     "@apollo/experimental-nextjs-app-support": "^0.10.1",
     "@apollo/react-hooks": "^4.0.0",
 
+    "valtio": "^1.13.2",    
+    "zustand": "^4.5.2",
+    "zustand-x": "^3.0.3",
+    "scheduler": "^0.23.2",
+
     "@auth0/auth0-react": "^2.2.4",
     
     "@emotion/cache": "^11.11.0",
@@ -62,12 +67,10 @@
     "@mui/utils": "^5.15.14",
     "@mui/x-data-grid": "^6.19.6",
     
-    "@react-spring/three": "^9.7.3",
-    "@react-spring/web": "^9.7.3",
     
     "@react-three/cannon": "^6.6.0",
-    "@react-three/drei": "^9.105.4",
-    "@react-three/fiber": "^8.16.2",
+    "@react-three/drei": "^9.105.6",
+    "@react-three/fiber": "^8.16.6",
     "@react-three/postprocessing": "^2.16.2",
     "@react-three/rapier": "^1.3.1",
     
@@ -103,12 +106,7 @@
     "use-count-up": "^3.0.1",
     
     "util": "^0.12.4",
-    "uuid": "^8.3.2",
-
-    "valtio": "^1.13.2",    
-    "zustand": "^4.5.2",
-    "zustand-x": "^3.0.3",
-    "scheduler": "^0.23.2"
+    "uuid": "^8.3.2"
 
   },
