+++ src/lib/threed/components/canvas/Canvas.jsx
@@ -2,18 +2,6 @@
 // ==========================================================
 // RESOURCES
 
-// ** APOLLO Imports
-import { useReactiveVar } from '@apollo/client'
-import {
-  isPreferencesSetVar,
-  preferencesDataVar,
-  preferencesStore,
-  projectStore,
-} from '#/lib/stores/apollo'
-
-// ** VALTIO ??
-import { proxy, useSnapshot } from 'valtio'
-
 import {
   Suspense,
   useState,
@@ -21,9 +9,6 @@ import {
   useRef,
   useTransition
 } from 'react'
-
-// ** LEVA CONTROLS Imports
-import { useControls } from 'leva'
 
 // THREE JS * ALL
 import * as THREE from 'three'
@@ -108,116 +93,6 @@ function ThreeDLoaderSimple() {
   return <Html center>THREED GUI LOADING... {Math.round(progress)} %</Html>
 }
 
-export function ThreeDEnvironment() {
-
-  // ** HOOKS
-  let prefs = useReactiveVar(preferencesDataVar)
-
-  // const [envPreset, setEnvPreset] = useState(prefs.environmentPreset) // 'park'
-  // You can use the 'inTransition' boolean to react to the loading in-between state,
-  // For instance by showing a message
-  const [inTransition, startTransition] = useTransition()
-
-  const [{ preset, blur }, setScenePreferencesLeva] = useControls(
-    'Scene Preferences',
-    () => (
-      {
-        preset: {
-          label: 'Environment',
-          value: prefs.environmentPreset, // envPreset,
-          // const presetsObj = {
-          //   apartment: 'lebombo_1k.hdr',
-          //   city: 'potsdamer_platz_1k.hdr',
-          //   dawn: 'kiara_1_dawn_1k.hdr',
-          //   forest: 'forest_slope_1k.hdr',
-          //   lobby: 'st_fagans_interior_1k.hdr',
-          //   night: 'dikhololo_night_1k.hdr',
-          //   park: 'rooitou_park_1k.hdr',
-          //   studio: 'studio_small_03_1k.hdr',
-          //   sunset: 'venice_sunset_1k.hdr',
-          //   warehouse: 'empty_warehouse_01_1k.hdr'
-          // };
-          // export { presetsObj };
-          options: [
-            'park', 'sunset', 'dawn', 'night', 'forest',
-            'studio', 'warehouse', 'apartment', 'lobby', 'city'
-          ],
-          // If onChange is present the value will not be reactive,
-          // see https://github.com/pmndrs/leva/blob/main/docs/advanced/controlled-inputs.md#onchange
-          // Instead we transition the preset value, which will prevents the suspense bound from triggering its fallback
-          // That way we can hang onto the current environment until the new one has finished loading ...
-          // onChange: (value) => startTransition(() => setEnvPreset(value))
-        },
-        blur: {
-          label: 'Bg Blur',
-          value: prefs.environmentBgBlur, // 0.00,
-          min: 0.00,
-          max: 0.20,
-        },
-      }
-    ),
-    {
-      color: 'darkgreen',
-      collapsed: false,
-      order: 0,
-    },
-  )
-
-  // ==========================================================
-  // ** environmentPreset
-  // **
-  useEffect(() => {
-    let newData = {...prefs}
-    // if (debug) console.debug('%c preset newData', ccm.green, newData)
-    newData.environmentPreset = preset
-    // if (debug) console.debug('%c preset newData UPDATED', ccm.green, newData)
-    preferencesDataVar(newData)
-    // if (debug) console.debug('%c preset preferencesDataVar', ccm.darkgreen, preferencesDataVar())
-    if (debug) console.debug('%c⚙️ READ FROM MASTER REACTIVE VAR: prefs.environmentPreset', ccm.yellowAlert, prefs.environmentPreset)
-  }, [preset])
-
-  // **
-  useEffect(() => {
-    // if (prefs.environmentPreset != undefined) {
-      setScenePreferencesLeva({ preset: prefs.environmentPreset })
-    // }
-    if (debug) console.debug('%c⚙️ READ FROM MASTER REACTIVE VAR: prefs.environmentPreset', ccm.greenAlert, prefs.environmentPreset)
-  }, [prefs.environmentPreset])
-
-  // ==========================================================
-  // ** environmentBgBlur
-  // **
-  useEffect(() => {
-    let newData = {...prefs}
-    // console.debug('%c blur newData', ccm.green, newData)
-    newData.environmentBgBlur = Math.round((blur + Number.EPSILON) * 100) / 100 // rounds to 2 decimal places
-    // console.debug('%c blur newData UPDATED', ccm.green, newData)
-    preferencesDataVar(newData)
-    // console.debug('%c blur preferencesDataVar', ccm.darkgreen, preferencesDataVar())
-    // setScenePreferencesLeva({ blur: blur })
-    if (debug) console.debug('%c⚙️ READ FROM MASTER REACTIVE VAR: prefs.environmentBgBlur', ccm.yellowAlert, preferencesDataVar().environmentBgBlur)
-  }, [blur])
-
-  // **
-  useEffect(() => {
-    // if (prefs.environmentBgBlur != undefined) {
-      setScenePreferencesLeva({ blur: prefs.environmentBgBlur })
-    // }
-    if (debug) console.debug('%c⚙️ READ FROM MASTER REACTIVE VAR: prefs.environmentBgBlur', ccm.greenAlert, prefs.environmentBgBlur)
-  }, [prefs.environmentBgBlur])
-
-  // ==========================================================
-
-  return (
-    <Environment
-      preset={prefs.environmentPreset}
-      blur={prefs.environmentBgBlur}
-      // blur={preferencesDataVar().environmentBgBlur}
-      background
-    />
-  )
-}
-
 // const controls = new OrbitControls(camera, renderer.domElement)
 // camera.lookAt(0.5, 0.5, 0.5)
 // controls.target.set(.5, .5, .5)
@@ -259,8 +134,8 @@ export function ThreeDCanvas({ _id, threeds }) { // , sceneState ??
   // if (debug) console.debug('%c📐 ThreeDCanvas props.threeds', ccm.darkredAlert, threeds)
   // if (debug) console.debug(`%c=======================================================`, ccm.darkred)
 
-  // ** HOOKS
-  const prefs = useReactiveVar(preferencesDataVar)
+  // // ** HOOKS
+  // const prefs = useReactiveVar(preferencesDataVar)
 
   // ** REF-erences using REACT
   const ref = useRef()
@@ -304,8 +179,9 @@ export function ThreeDCanvas({ _id, threeds }) { // , sceneState ??
       >
   
         {/* @ts-ignore */}
-        <threedIO.Out />
-        <Preload all />
+        {/* <threedIO.Out /> */}
+
+        {/* <Preload all /> */}
 
         {/* SUSPENSEFUL... */}
         {/* <Suspense fallback={<Html>HEY HEY HEY</Html>}> */}
@@ -345,52 +221,3 @@ export function ThreeDCanvas({ _id, threeds }) { // , sceneState ??
 }
 
 export default ThreeDCanvas
-
-// ** NOTES
-// **
-// function Watch(props) {
-//   const ref = useRef()
-//   const { nodes, materials } = useGLTF('/objects/examples/watch-v1.glb')
-//   useFrame((state) => {
-//     const t = state.clock.getElapsedTime()
-//     ref.current.rotation.x = -Math.PI / 1.75 + Math.cos(t / 4) / 8
-//     ref.current.rotation.y = Math.sin(t / 4) / 8
-//     ref.current.rotation.z = (1 + Math.sin(t / 1.5)) / 20
-//     ref.current.position.y = (1 + Math.sin(t / 1.5)) / 10
-//   })
-//   return (
-//     <group ref={ref} {...props} dispose={null}>
-//       <mesh geometry={nodes.Object005_glass_0.geometry} material={materials.glass}>
-//         <Html scale={100} rotation={[Math.PI / 2, 0, 0]} position={[180, -350, 50]} transform occlude>
-//           <div className='annotation'>
-//             <span style={{ fontSize: '1.5em' }}>🥕</span> 6,550
-//           </div>
-//         </Html>
-//       </mesh>
-//       <mesh castShadow receiveShadow geometry={nodes.Object006_watch_0.geometry} material={materials.watch} />
-//     </group>
-//   )
-// }
-
-// function CoffeeCup(props) {
-//   const { nodes, materials } = useGLTF('/objects/examples/coffee-transformed.glb')
-//   console.debug('materials', materials)
-//   return (
-//     <mesh
-//       receiveShadow
-//       castShadow
-//       geometry={nodes.coffee_cup_top_16oz.geometry}
-//       material={materials['13 - Default']}
-//       {...props}
-//       dispose={null}
-//     />
-//   )
-// }
-
-// // EXAMPLE ANIMATION using hook 'useFrame' (with 'useRef' references)
-// function ActionRig() {
-//   return useFrame((state) => {
-//     state.camera.position.x = THREE.MathUtils.lerp(state.camera.position.x, 1 + state.mouse.x / 4, 0.075)
-//     state.camera.position.y = THREE.MathUtils.lerp(state.camera.position.y, 1.5 + state.mouse.y / 4, 0.075)
-//   })
-// }
